pyproject.toml
[project]
name = "moviedb"
version = "1.0.1"
description = "AI-powered movie recommendation application"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-admin-numeric-filter>=0.1.9",
    "django-environ>=0.12.0",
    "ipython>=9.7.0",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "torch>=2.9.1",
    "streamlit>=1.51.0",
    "openai>=2.8.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
